Update to Scalang 0.17 and adjust metrics to match
The instrumentedName method is no longer available. While we could
switch to using self.toErlangString directly, I've elected to send all
search metrics to the same place, since we a) always want the sum
across all index service processes and b) it obviates the need to
cleanup MBeans.

The upgrade to Scalang 0.17 includes an important fix for process
monitoring during initialization as well as the automatic removal of
MBeans when processes exit. It also incorporates a few small
enhancements from me, one for small atoms support and another for
handling NaN/Infinity/-Infinity values. The latter will allow cleaner
support for sorting (because score is not calculated and is therefore
NaN).

Note: The metrics names have changed as a result of these changes.