Use two-phase group search
Lucene grouping is naturally two phase, the first to get the ordered
list of 'top' groups, the second to get the top documents within those
groups. I had hoped to get away with one phase but the anomalies were
too great (each shard returning a different order of top groups and,
by omitting the first phase, I don't have the information to stabilize
the order).

This is a big patchset that conflates a few things. Notably, the
message format has changed in addition to the switch to handling each
phase separately. The end result is cleaner than the diff.