Don't add a facet containing the default delimiter character
Any name or value of a facet that contains U+1F (a control character)
causes Clouseau to crash, leading to wedging as the crash occurs where
it kills an actor/service where it won't be restarted.

Since users should not expect control characters to work inside facet
values, we simply omit to add them.

BugzID: 43209