Update LRU when index is used
Prior to this commit only index opening/lookup would update the
LRU. This commit updates the LRU on index activity, as it should
always have done. The effect is that we will close the least-recently
used index from now on.

BugzID: 75594

Ported from https://github.com/cloudant-labs/clouseau/pull/3 where it was +1'd.