Update pom.xml to use the absolute path for clouseau.dir config
Currently we are using the relative path for setting the clouseau.dir in
dev launcher configurations. As a result it would resolve the path to the
working directory specified by the JVM property user.dir. Change this to
absolute path and use the basedir config of the maven, which will be
resolved to the same directory as pom.xml.

BugzId:87336