Change thread model again
We need to ensure that at most one thread is processing the messages
for a Service at a time, which previous versions of this failed at.

I think the `lock` in ServiceRegistry is probably highly contended and
a likely bottleneck, so much room for improvement.

With this change, we can run concurrent queries to indexes of q=1 and
q>1, so that's nice.