Solidify error handling
This patch ensures that all services exit by throwing OtpErlangExit so
that all causes of exit are handled correctly and in the same manner,
whether the reason was external (the linked dreyfus process exiting,
for example) or internal (an I/O exception at the disk level).

This necessarily forced a rewrite of how asynchronous index commit was
done, to ensure that exits are not caused from the wrong thread. The
implementation now closely matches clouseau's.