Remove metrics on field cache size
The field cache metrics do not provide much value and if one wants to move a
later version of Lucene, it gets more complicated to get this right.  According
to @rnewson, this metrics was introduced to investigate a customer issue back
in the past, but it proved unless even for that.