Fix race condition between closing the index and cleaning it up
When a database is deleted, the corresponding directory structure
is removed entirely, even if the index writer has not yet been
closed.  Due to the missing files, both attemps of closing the
index and then rolling it back fail, which eventually leads to
leak of resources, more specificially locks, due to a known bug in
Lucene versions prior to 4.7.1 [1][2].  In result, the subsequent
search requests to the same index will fail when recreated.

The reason for this is that the clean-up service might be called
earlier than killing off the index service and the order of removal
(first the lock, then the rest of the files) may become reversed.
Adjust the services to handle the situation gracefully.

[1] https://issues.apache.org/jira/browse/LUCENE-5544
[2] But is not a good practice to delete directories while being
    used, anyway.