Add metric on the count of NativeFSLocks held
Lucene 4.6 is known to leak locks when an exception hits the rollback of an
index writer.  Introduce an optional metric on the size of the internal store
that tracks the currently held NativeFSLocks to be able to assess its severity
in practice.

The method is specific to the Lucene internals and it uses Java reflection to
extract the data as it is officially not accessible.