Avoid matching on `case Some(fields: List[(String, Any)])`
The matching on inner types of the container will not work on next version of
Scala. Newer Scala causes type erasure when matching on inner type of the
container (such as List in this case).

This refactoring also avoid using index based access to tuple elements and uses
pattern matching instead which improves readability.

Assumptions

1. The `fields` value is a list
2. The elements of a `fields` list are tuples `(String, String)`
3. It is ok to skip elements which do not match the `(String, String)` pattern