Introduce types to supported analyzers take2 (#84)
* Refactor createAnalyzer to remove polymorphic arguments

The polymorphism makes it harder to introduce types. Because it requires union
types. Which are not natively supported by the version of Scala we use.
Elimination of the polymorphism would allow us to introduce types.

The introduction of types is the vehicle to solve type erasure problem we would
have to deal with when we upgrade Scala to the next version.

The refactoring is done using following steps

1. Introduce `AnalyzerOptions` class with type specific constructors
  - `def fromMap(map: Map[_, _])`
  - `def fromAnalyzerName(name: String)`
  - `def fromKVsList(options: List[_])`
2. Make sure we correctly go from `Any` to the concrete type. This PR uses the
  `.collect` combinator instead of relying on `ClassCastException`.

Assumptions

1. The keys of `options` passed to `OpenIndexMsg` are strings.
2. It is ok to just ignore all non-string keys in `options`.
3. The analyzer name is either a `String` or a single `String` element wrapped
   in the `List`.
4. The keys of options passed to OpenIndexMsg are strings.
5. It is ok to just ignore all non-string keys in options.
6. The fields value is a list.
7. The elements of a fields list are tuples.
   `(String, String)` or (String, [String]).
8. The config is a String in ('analyze, config, text) message in
   `AnalyzerService.handleCall` and it should really be named
   `('analyze, analyzerName, text)`.

* Avoid matching on `Some(stopwords: List[String])`

The matching on inner types of the container will not work on next version of
Scala. Newer Scala causes type erasure when matching on inner type of the
container (such as List in this case).

Assumptions

1. The `stopwords` is a list
2. The elements of a `stopwords` are strings.
3. It is ok to skip non-strings elements of `stopwords`.