scalang: implement `Service#exit` and handle remote exits
Through the use of `MessageEnvelop.Exit`, make it possible for
processes to initiate their own shutdown by `OTPNode#stopActor`.
The respective queues (from reading local and remote messages) are
stopped on receiving this type of message, which allows the actors
to terminate cleanly.

When the connection breaks or a remote process sends an exit signal
(see `erlang:exit/2` [1]), an `OtpErlangExit` exception is thrown
by `OtpMbox#receiveMsg` and that is converted to a
`MessageEnvelope.Exit` message so that the process could also stop
in response to that event.  Trapping exit signals is not yet
supported.

[1] https://www.erlang.org/doc/man/erlang.html#exit-2