Allow override of location of Rebar
Although the name of the Rebar binary is factored out as a shell
variable, there is no way to modify that.  Tweak the definition a
bit to allow users to change it externally so the build process
could be fit with both Rebar 2.x and 3.x, as needed.

It could be used as follows, for example:

```console
REBAR=rebar3 make zeunit
```