Remove the finalizer on `AddressableActor#stream`
In `NodeProcess#loop`, the stream is already closed as part of
shutting down the context on receiving an "exit" message.  The
`stopActor` method is responsible for handling the exit and
calling the actor's `onTermination` method.  The stream's finalizer
implemented a similar functionality but this has become redundant by
now and it mistakenly called `onTermination` once more on the
resulting interruption.