Restructure processes to have separate `init` and `echo`
Rename the `coordinator` process to `init` and move the echo
functionality to a dedicated process family, `echo`.  The instances
of `echo` could be then spawned via `init` on request.  Besides the
better code structure, it could also help with testing process
creation and destruction as well as with fan-out scenarios.