OTP mailbox notifications
Extend the OTP mailbox delivery logic in the Java Interface to
notify listeners on the arrival of each new message through a
callback mechanism.  With the help of that, ZIOSE could be now
changed to read the messages immediately after they have been
placed in the queue of each mailboxes.  This requires less active
IO threads in the run-time system and helps with scalability.