ziose: implement creation of Erlang references
Use the Erlang/OTP Java Interface to fill in the stub for the
definition of the `makeRef` function.  This provides better
guarantees for staying in synch with the upstream implementation.
In parallel to that, move the definition over to `Node` / `OTPNode`
as this method is not tied to mailboxes but the nodes themselves.

The node reference is embedded in the environment of ZIO actions,
which makes it possible to use `makeRef` there as follows:

```scala
for {
  node <- ZIO.service[Node]
  ref <- node.makeRef()
  ...
} yield ...
```