Add `erlfmt` for checking/applying formatting of Erlang sources
The `erlfmt` tool [1] is actively used in the Apache CouchDB project
to keep track of formatting issues in the code base.  With the help
of the `erlfmt-format` target, `erlfmt` could be used to
automatically reformat the sources when needed.

Note that `erlfmt` does not have its own nice packaging, therefore
an extension of `scripts/tools.sh` was required to handle
installing programs from source, using a `git` clone of their home
repository.

[1] https://github.com/WhatsApp/erlfmt