Implement creation of monitors and links
Complement the existing support for monitors with the capability of
creating monitors and links from the Scala side, rooted in the
Erlang/OTP Java interface layer.  This obviously implies further
modifications to the Java Interface, mostly for the sake of
simplicity, that could be later submitted to upstream.

Note that since the receiver logic of monitors and all the links
have already been implemented in Java, this work does not yet
attempt to loop back local-to-local monitors or links in the Scala
layer, all such requests will be passed through the Java layer at
the moment.  A clear benefit of this approach is that monitors or
links could be created for remote nodes as well.