Fix LogHistory.withContextAnnotation matchers
There was a problem with withContextAnnotation based matchers.
In some cases they where returning things which do not suppose to match.
Specifically the idea bihing the API of withContextAnnotation was to include
the event into result set only iff both conditions satisfy:

- the provided partial function is defined on the log event
- the provided partial function return true

However the existing implementation didn't use the result of the invocation at all.
This PR fixes this problem by. Returning the boolean result of an invocation.