Structured logging for standard output
In modern deployment environments, log lines with a JSON-based
structure could be integrated the best way, thus implement support
for that.  This can be nicely done through `consoleJsonLogger`
from `zio.logging` [1].  Note that the `micrometer` library on
which JMX support is based, pulls in SLF4J, therefore a bridge is
added to translate the messages from that layer for uniformity.
For completeness, translate all the `ZIO.debug` calls to ones that
are made to a logger instance so those messages are going to be
handled accordingly.

The format of the logs is as follows (pretty printed):

```console
{
  "timestamp": "2024-09-03T11:52:35.94446+02:00",
  "level": "DEBUG",
  "fiberId": "zio-fiber-1307834791",
  "annotations": {
    "logger_name": "clouseau.InitServiceBuilder"
  },
  "location": {
    "module": "InitService.scala",
    "line": 132
  },
  "message": "Started init"
}
```

That is not enabled by default though because it is not so
human-readable which makes it harder to employ for local
development.  To switch over to the structured format, one has
to set the `logger.output` configuration parameter to `JSON`.
Otherwise the emitted log events are going to be formatted
like this:

```console
2024-09-03T09:52:35.945Z DEBUG  zio-fiber-1307834791  InitService.scala:132    [clouseau.InitServiceBuilder] Started init
```

[1] https://zio.dev/zio-logging/console-logger/