depcheck: suppress CVE-2024-45772
This vulnerability [1] does not apply to our use case of Lucene.  It
is rooted in the Internet-facing components, the HTTP interface of
the replicator, which is not used here.  Unfortunately, the CPE 2.3
definition does not make a difference between the unaffected and
affected components and marks every Lucene module vulnerable.

[1] https://nvd.nist.gov/vuln/detail/CVE-2024-45772