Output printable termination reason from ClouseauSupervisor
The optimization which prints Ebinary as string doesn't work when the message
is received via jInterface. Which is the case for `handleMonitorExit` used in
ClouseauSupervisor. It is very inneficient to determine printability of
arbitrary strings. This PR adds a special case when message starts from `Exception`
prefix. This should cover 100% of messages received by ClouseauSupervisor.

I also added `(implicit: Trace)` to some callbacks in `Service` to improve
accuracy of stack traces.