Fix reading `clouseau.dir` from the configuration file
The root directory for the search indexes cannot be read from the
configuration file.  It is translated to a value of the `RootDir`
type but the recorded value is not extracted from that on querying
it from the Clouseau code base as `clouseau.dir` but the internal,
hard-wired default is returned unconditonally always.

Curiously, the unit tests do not assume the `RootDir:` wrapper
around the value in the sample configuration file, so there must be
some discrepancy here.

Mitigate the issue by rewriting the query routine to extract the
captured value when it is specified.