Restore the `clouseau.count_locks` configuration flag
The `count_locks` Boolean option is used in Clouseau to enable
tracking the number of locks in Apache Lucene and tell if they are
piling up which can signify leaking.

For what it is worth, since we backported fixes for Lucene 4.6.1 to
address that, it has lost its actual relevance and it shall be
removed as such.  But it might still serve as a way to test if
gauge metrics are working as expected.