Quote reported JMX metrics
The `collectd` plugin that collects the JMX metrics for Clouseau
assumes that the `com.cloudant.clouseau` family of metrics are all
quoted, otherwise it will not be able to find those.  (Well, it has
some leftover support for reading the values without quotes, but it
crashes currently due to inconsistencies in the code.)

Adjust the metrics generation to fix this issue.