Fix LRU inconsistency
Returning true from removeEldestEntry means pathToPid has one fewer element than
pidToPath until trapMonitorExit is called in IndexManagerService for that
element.

Change how we enforce the upper limit of the LRU so that the two maps stay in
sync, and add assertions on size to prove it.

This may fix some of the runtime issues around 'noproc'