logging: introduce Syslog support
In its current form, ZIOSE is capable of generating log events to the
standard output.  This might be enough for deployments where a
client-side `syslog` daemon could be independently configured to read
the resulting log file and forward the events.  However, this carries
the risk of putting I/O pressure on the disk, and it would be nice to
let ZIOSE communicate directly to a `syslog` daemon over the network
(TCP or UDP) to mitigate such scenarios, if necessary.

The Syslog output could be configured as follows, for example.

```
logger: {
  format: JSON
  output: Syslog
  syslog: {
    protocol: UDP      # UDP or TCP, default: UDP
    host: 127.0.0.1    # default: localhost
    port: 514          # default: 514 (standard Syslog port)
    facility: LOCAL5   # default: CONSOLE
    level: info        # default: debug
    tag: "clouseau: "  # default: <empty>
  }
}
```

Note that this commit creates a breaking change for the application
configuration as the semantics of `logger.output` has been revisited.
Previously, `logger.output` might be either `PlainText` or `JSON`,
which would be now stored under `logger.format`.  The configuration
value `Syslog` could be used for  `logger.output` to create the
connection for Syslog, and `Stdout`, the default value, could be
used to maintain the previous behavior, that is, sending events to
the standard output.