Introduce the `Raw` log format
The original `PlainText` output format includes ANSI escape codes
to colorize the output as well as it prints the messages tabulated,
much more optimized for human consumption.  This may not be
suitable for regular logging purposes because the color codes shall
be removed for parsing and the padding makes the text longer, which
increases the required storage size.

As an alternative between fancy plain-text and compact
JSON-formatted log messages, add another one that combines both:
another plain-text format which loses the color codes and the extra
padding.

This new format now looks like this:

```console
2025-04-03T21:04:02.016Z [DEBUG][zio-fiber-1333081487][ClouseauSupervisor.scala:163][clouseau.supervisor] main is started after 12 ms
```

In accordance with this extension, shorten the name `PlainText` to
`Text` to make it somewhat easier to use.  Also, make `Raw` the
default format to stay on the conservative side.