Eliminate `println` for logging, improve log event rendering
There were a couple of `println` instance left in the source that
should be now translated to regular logging actions to avoid losing
them on processing.  While here, create a custom `toString`
implementation for the `Exponent` type to make it easier to
understand when values of that type are printed.