Fix Erlang cookie handling in tests
One must be careful with setting the `cookie` value for the
application, because after the naming the JAR file, it expects one
listing the parameters for the application itself.  Hence any
overrides to the configuration must be provided before that.  That
is, in case of the Erlang cookie:

```console
$ java -Dcookie=COOKIE -jar jarfile.jar
```

Also, CouchDB upstream has recently gained full support for passing
around the Erlang cookie [1], so leverage that here too.

[1] https://github.com/apache/couchdb/pull/5531