Temporary fix: Prevent CouchDB from hanging when running mango-tests
Starting a ziose node and immediately running the concurrent health
check causes CouchDB to fail to check if the cluster is up and thus
fail to run Mango tests. Even though the ziose node passes the health
check, some underlying components are not working properly. CouchDB
fails to start and gets stuck on `dev/run script` (`urlopen(url)`).
Since Mango tests are based on the `requests` library, some tests
will also get stuck if start Mango tests in another terminal.

This behavior only occurs in Linux and not in MacOS. Therefore, add
a time interval and `retry(3)` to temporarily fix this issue.