Make externalMailbox unbounded
The `jInterface` has a single threaded implementation. When `jInterface` calls
the `onMessageReceived` it expects immediate return. OTherwise the `jInterface`
wouldn't be handling next requests until the `onMessageReceived` unblocks.

This behaviour of `jInterface` contradicts with idea of a backreassure. The idea
was that when the number of messages in the actor's mailbox reacheas the
configured capacity we would block the process which calls `mailbox.offer(_)`.

If we cannot block inside the `jInterface` we only left with two options.
Change the implementation of `jInterface` to support blocking of individual
actors or remove backpreassure from the queue which use inside the `jInterface`.

This commit makes the `externalMailbox` unbounded so we wouldn't block the
`jInterface`'s main loop. While keeping the backpreasure for internal messages.

The test case was updated to cause the blocking behaviour by changing amount of
messages we send to test actor to be greater than configured capacity.