Unbreak dependency scanning by upgrading to version 12.1.6
Sonatype OSS Index has recently introduced authentication for the
use of their API [1], for which we do not yet have an API key
ready.  This is a common problem to many CI/CD systems all around
the world, thus version 12.1.6 of OWASP Dependency Check [2] was
published where the use of OSS Index is automatically disabled
when no credentials are provided.  `sbt-dependency-check` 1.8.2
[3] was also published to chase this release, so move to that
version and fix our CI.

Note that it is going to lead to a potential loss of coverage,
but at least it keeps working.

[1] https://ossindex.sonatype.org/doc/auth-required
[2] https://github.com/dependency-check/DependencyCheck/blob/main/CHANGELOG.md#version-1216-2025-09-24
[3] https://github.com/nMoncho/sbt-dependency-check/releases/tag/v1.8.2