ci: make `timeout` configurable and run in the foreground
On macOS, the `timeout` command that is shipped with `coreutils`
(GNU port) tends to get stuck for running `mango-test`.  Possibly
that is not a coincidence that there exists a native version [1]
which is recommended for use instead.

Make it possible for the developers to override the use of that via
factoring it out into a `make` variable.  While there, pass the
`--foreground` flag to work around the issue by forbidding the tool
going to the background.

[1] https://github.com/denispol/darwin-timeout