Restore object form of fields in perfield analyzer definition
An analyzer, even one nested inside the "default" or "fields" section of a
"perfield" analyzer can be expressed as an object.

Without this fix indexes have silently used StandardAnalyzer when users have
declared other analyzers.

This significant regression was introduced in
https://github.com/cloudant-labs/clouseau/pull/84 and specifically this false
assumption "The analyzer name is either a String or a single String element
wrapped in the List."